Extract review job module
